requirements.txt
pylint
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
requests
gunicorn
